backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
ast_serialize==0.3.0
attrs==26.1.0
babel==2.18.0
bcrypt==4.1.3
beautifulsoup4==4.14.3
black==26.3.1
boto3==1.43.5
botocore==1.43.5
build==1.5.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.9
click==8.3.3
courlan==1.3.2
cryptography==48.0.0
cuda-bindings==13.2.0
cuda-pathfinder==1.5.4
cuda-toolkit==13.0.2
dateparser==1.4.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.29.0
flake8==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.51.0
google-auth-httplib2==0.4.0
google-genai==1.75.0
google-generativeai==0.8.6
googleapis-common-protos==1.75.0
greenlet==3.5.0
groq==1.2.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.5.0
htmldate==1.9.4
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.14.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==7.1.0
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
jq==1.11.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
kubernetes==35.0.0
librt==0.10.0
litellm==1.80.0
lxml==6.1.0
lxml_html_clean==0.4.4
markdown-it-py==4.1.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.1
motor==3.3.1
mpmath==1.3.0
multidict==6.7.1
mypy==2.0.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-nvjitlink==13.0.88
nvidia-nvtx==13.0.85
oauthlib==3.3.1
onnxruntime==1.25.1
openai==1.99.9
opentelemetry-api==1.41.1
opentelemetry-exporter-otlp-proto-common==1.41.1
opentelemetry-exporter-otlp-proto-grpc==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
orjson==3.11.9
overrides==7.7.0
packaging==26.2
pandas==3.0.2
passlib==1.7.4
pathspec==1.1.1
pillow==12.2.0
platformdirs==4.9.6
playwright==1.59.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.28.0
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.14.1
pydantic_core==2.46.4
pyee==13.0.1
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.5.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.27
pytokens==0.4.1
pytz==2026.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.17.0
s5cmd==0.2.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.37.2
stripe==15.1.0
sympy==1.13.1
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tld==0.13.2
tokenizers==0.22.2
torch==2.5.1
tqdm==4.67.3
trafilatura==2.0.0
transformers==5.8.0
typer==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.25.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
yarl==1.23.0
zipp==3.23.1
